move TestCase and Port to public class
