+++ AraigTestFrameworkGenerator/de.fraunhofer.ipa.ros.araig.plugin/src/de/fraunhofer/ipa/ros/araig/plugin/GenerationHandler.java
@@ -52,77 +52,6 @@ import de.fraunhofer.ipa.ros.araig.plugin.generator.AMBSGenerator;
 import de.fraunhofer.ipa.ros.araig.plugin.generator.CustomOutputProvider;
 import rossystem.RosSystem;
 
-class Port {
-	private String name = "";
-	private String msg_type  = "";
-	private String description  = "";
-	
-	public String getName() {
-        return name;
-    }
-    public void setTopic(String name) {
-        this.name = name;
-    }
-	public String getMsgType() {
-        return msg_type;
-    }
-    public void setMsgType(String msg_type) {
-        this.msg_type = msg_type;
-    }
-    
-    public String getDescription() {
-        return description;
-    }
-    public void setDescription(String description) {
-        this.description = description;
-    }
-	@Override
-	public String toString() {
-		return "{name: "+ name
-				+ "\n msg type: " + msg_type 
-				+ "\n description: " + description
-				+"\n }";
-	}
-}
-
-class TestCase {
-	private String name = "";
-	private String description = "";
-	private Port[] subscribers;
-	private Port[] publishers;
-	private Port[] action_clients;
-	
-	public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
-	
-    public String getDescription() {
-        return description;
-    }
-    public void setDescription(String description) {
-        this.description = description;
-    }
-    public Port[] getPublishers() {
-        return publishers;
-    }
-    public Port[] getSubscribers() {
-        return subscribers;
-    }
-    public Port[] getActionClients() {
-        return action_clients;
-    }
-    @Override
-    public String toString() {
-        return name + ":{ \n publishers:" + Arrays.toString(publishers) 
-        			+ "\n subscribers:" + Arrays.toString(subscribers) 
-        			+ "\n action_clients:" + Arrays.toString(action_clients)
-        			+ "\n}";
-    }
-}
-
 public class GenerationHandler extends AbstractHandler implements IHandler {
 	
 	@Inject
